+++ include/tomop/tomop.hpp
@@ -9,3 +9,4 @@
 #include "packets/partitioning_packets.hpp"
 #include "packets/reconstruction_packets.hpp"
 #include "packets/scene_management_packets.hpp"
+#include "packets/control_packets.hpp"